move Python 3.12 specific tests to test_union_as_type_alias_recursive.py
