+++ tests/unit/v1/test_loaders.py
@@ -979,39 +979,6 @@ def test_union_recursive():
     )
 
 
-# noinspection PyCompatibility
-@pytest.mark.skipif(not PY312_OR_ABOVE, reason='Requires Python 3.12 or higher')
-def test_union_as_type_alias_recursive():
-    """
-    Recursive or self-referential `Union` (defined as `TypeAlias`)
-    types are supported.
-    """
-    type JSON = str | int | float | bool | dict[str, JSON] | list[JSON] | None
-
-    @dataclass
-    class MyTestClass(JSONWizard):
-
-        class _(JSONWizard.Meta):
-            v1 = True
-
-        name: str
-        meta: str
-        msg: JSON
-
-    x = MyTestClass.from_dict(
-        {
-            "name": "name",
-            "meta": "meta",
-            "msg": [{"x": {"x": [{"x": ["x", 1, 1.0, True, None]}]}}],
-        }
-    )
-    assert x == MyTestClass(
-        name="name",
-        meta="meta",
-        msg=[{"x": {"x": [{"x": ["x", 1, 1.0, True, None]}]}}],
-    )
-
-
 def test_multiple_union():
     """Test case for a dataclass with multiple `Union` fields."""
